+++ gen/strings.go
@@ -31,8 +31,8 @@ func RuneNoControl() gopter.Gen {
 func genRune(int64Gen gopter.Gen) gopter.Gen {
 	return int64Gen.Map(func(value int64) rune {
 		return rune(value)
-	}).SuchThat(func(v interface{}) bool {
-		return utf8.ValidRune(v.(rune))
+	}).SuchThat(func(v rune) bool {
+		return utf8.ValidRune(v)
 	})
 }
 
@@ -94,8 +94,7 @@ func Identifier() gopter.Gen {
 		tail := values[1].([]rune)
 		result := make([]rune, 0, len(tail)+1)
 		return string(append(append(result, first), tail...))
-	}).SuchThat(func(v interface{}) bool {
-		str := v.(string)
+	}).SuchThat(func(str string) bool {
 		if len(str) < 1 || !unicode.IsLower(([]rune(str))[0]) {
 			return false
 		}
@@ -109,8 +108,8 @@ func Identifier() gopter.Gen {
 }
 
 func genString(runeGen gopter.Gen, runeSieve func(ch rune) bool) gopter.Gen {
-	return SliceOf(runeGen).Map(runesToString).SuchThat(func(v interface{}) bool {
-		for _, ch := range v.(string) {
+	return SliceOf(runeGen).Map(runesToString).SuchThat(func(v string) bool {
+		for _, ch := range v {
 			if !runeSieve(ch) {
 				return false
 			}