+++ CHANGELOG.md
@@ -18,6 +18,22 @@
      return strings.ToUpper(v)
   })
   ```
+- Correspondingly Gen.SuchThat now also ccept `interface{}` instead of `func (interface{}) bool`
+
+  This allows cleaner sieve functions without type conversion. E.g. instead of
+
+  ```Go
+  gen.AnyString().SuchThat(function (v interface{}) bool {
+     return HasPrefix(v.(string), "P")
+  })
+  ```
+  you can (and should) now write
+
+  ```Go
+  gen.AnyString().SuchThat(function (v string) bool {
+     return HasPrefix(v, "P")
+  })
+  ```
 - Gen.FlatMap now has a second parameter `resultType reflect.Type` defining the result type of the mapped generator
 - Reason for these changes: The original `Map` and `FlatMap` had a recurring issue with empty results. If the original generator created an empty result there was no clean way to determine the result type of the mapped generator. The new version fixes this by extracting the return type of the mapping functions.